Custom parsing instead of bincode
Switch from `bincode` to custom parsing macros for Flir data
structs.  Rationales for the switch:

1. Technically, `bincode` doesn't provide a concrete spec
for how it serializes structs. While it does match the usual
packed encoding as in C, and allows changing endianness,
this is not guaranteed to be stable.

2. While `bincode` provides endianness, the FLIR and other
image formats seem to like changing endianness within the
code based on flags in the header.  This makes it a bit
roundabout to use the `serde` + `bincode` as the endianness
is set in the outer context.

3. More flexible error reporting, and one-off tuple decoding
is easier with our custom macros.

** Code Changes

+ parse macros for parsing packed structs
+ replace bincode dependency with custom parsing code